Added RakeCommit namespace in order to avoid conflicts with Git gem in Rails 3.1
