+++ bin/rake_commit
@@ -1,5 +1,5 @@
 #!/usr/bin/env ruby
 
-require File.expand_path(File.dirname(__FILE__) + "/../lib/commit")
+require File.expand_path(File.dirname(__FILE__) + "/../lib/rake_commit")
 
-Commit.new.commit
+RakeCommit::Commit.new.commit
