+++ lib/rake_commit/cruise_status.rb
@@ -1,0 +1,27 @@
+require 'rexml/document'
+require "open-uri"
+
+module RakeCommit
+  class CruiseStatus
+
+    def initialize(feed_url)
+      project_feed = open(feed_url).read
+      @doc = REXML::Document.new(project_feed)
+    rescue Exception => e
+      @failures = [e.message]
+      @doc = REXML::Document.new("")
+    end
+
+    def pass?
+      failures.empty?
+    end
+
+    def failures
+      @failures ||= REXML::XPath.match(@doc, "//item/title").select { |element|
+        element.text =~ /failed$/
+      }.map do |element|
+        element.text.gsub( /(.*) build (.+) failed$/, '\1' )
+      end
+    end
+  end
+end
