+++ test/integration_test.rb
@@ -16,19 +16,19 @@ class IntegrationTest < Test::Unit::TestCase
     Dir.chdir(TMP_DIR) do
       FileUtils.mkdir "git_repo"
       Dir.chdir("git_repo") do
-        Shell.system "echo 'task :default do; end' >> Rakefile"
+        RakeCommit::Shell.system "echo 'task :default do; end' >> Rakefile"
         create_git_repo
       end
 
       in_git_repo do
-        Shell.system "touch new_file"
-        Shell.system "yes | ../../../bin/rake_commit"
+        RakeCommit::Shell.system "touch new_file"
+        RakeCommit::Shell.system "yes | ../../../bin/rake_commit"
 
-        log_lines = Shell.backtick("git log --pretty=oneline").split("\n")
+        log_lines = RakeCommit::Shell.backtick("git log --pretty=oneline").split("\n")
         assert_equal 2, log_lines.size
         assert_match /y - y/, log_lines.first
 
-        assert_equal "", Shell.backtick("git cherry origin")
+        assert_equal "", RakeCommit::Shell.backtick("git cherry origin")
       end
     end
   end
@@ -37,26 +37,26 @@ class IntegrationTest < Test::Unit::TestCase
     Dir.chdir(TMP_DIR) do
       FileUtils.mkdir "git_repo"
       Dir.chdir("git_repo") do
-        Shell.system "echo 'task :default do; raise \"failing test\"; end' >> Rakefile"
+        RakeCommit::Shell.system "echo 'task :default do; raise \"failing test\"; end' >> Rakefile"
         create_git_repo
       end
 
       in_git_repo do
-        Shell.system "touch new_file"
+        RakeCommit::Shell.system "touch new_file"
 
         begin
-          Shell.system "yes | ../../../bin/rake_commit"
+          RakeCommit::Shell.system "yes | ../../../bin/rake_commit"
           fail
         rescue => e
         end
 
-        log_lines = Shell.backtick("git log --pretty=oneline").split("\n")
+        log_lines = RakeCommit::Shell.backtick("git log --pretty=oneline").split("\n")
         assert_equal 2, log_lines.size
         assert_match /y - y/, log_lines.first
 
         unpushed_sha = log_lines.first.gsub(/ .*/, "")
 
-        assert_equal "+ #{unpushed_sha}\n", Shell.backtick("git cherry origin")
+        assert_equal "+ #{unpushed_sha}\n", RakeCommit::Shell.backtick("git cherry origin")
       end
     end
   end
@@ -65,14 +65,14 @@ class IntegrationTest < Test::Unit::TestCase
     Dir.chdir(TMP_DIR) do
       FileUtils.mkdir "git_repo"
       Dir.chdir("git_repo") do
-        Shell.system "echo 'task :default do; end' >> Rakefile"
+        RakeCommit::Shell.system "echo 'task :default do; end' >> Rakefile"
         create_git_repo
       end
 
       in_git_repo do
-        Shell.system "yes | ../../../bin/rake_commit"
+        RakeCommit::Shell.system "yes | ../../../bin/rake_commit"
 
-        log_lines = Shell.backtick("git log --pretty=oneline").split("\n")
+        log_lines = RakeCommit::Shell.backtick("git log --pretty=oneline").split("\n")
         assert_equal 1, log_lines.size
         assert_match /Added Rakefile/, log_lines.first
       end
@@ -83,22 +83,22 @@ class IntegrationTest < Test::Unit::TestCase
     Dir.chdir(TMP_DIR) do
       FileUtils.mkdir "git_repo"
       Dir.chdir("git_repo") do
-        Shell.system "echo 'task :default do; end' >> Rakefile"
+        RakeCommit::Shell.system "echo 'task :default do; end' >> Rakefile"
         create_git_repo
       end
 
       in_git_repo do
-        Shell.system "git checkout -b br"
-        Shell.system "echo 'blah' >> one"
-        Shell.system "git add one"
-        Shell.system "git commit -m 'commit on branch'"
-        Shell.system "git checkout master"
-        Shell.system "git merge --no-ff br"
+        RakeCommit::Shell.system "git checkout -b br"
+        RakeCommit::Shell.system "echo 'blah' >> one"
+        RakeCommit::Shell.system "git add one"
+        RakeCommit::Shell.system "git commit -m 'commit on branch'"
+        RakeCommit::Shell.system "git checkout master"
+        RakeCommit::Shell.system "git merge --no-ff br"
 
-        assert_equal 3, Shell.backtick("git log --pretty=oneline").split("\n").size
-        Shell.system "yes | ../../../bin/rake_commit"
+        assert_equal 3, RakeCommit::Shell.backtick("git log --pretty=oneline").split("\n").size
+        RakeCommit::Shell.system "yes | ../../../bin/rake_commit"
 
-        log_lines = Shell.backtick("git log --pretty=oneline").split("\n")
+        log_lines = RakeCommit::Shell.backtick("git log --pretty=oneline").split("\n")
         assert_equal 2, log_lines.size
         assert_match /y - y/, log_lines.first
         assert_match /Added Rakefile/, log_lines.last
@@ -110,22 +110,22 @@ class IntegrationTest < Test::Unit::TestCase
     Dir.chdir(TMP_DIR) do
       FileUtils.mkdir "git_repo"
       Dir.chdir("git_repo") do
-        Shell.system "echo 'task :default do; end' >> Rakefile"
+        RakeCommit::Shell.system "echo 'task :default do; end' >> Rakefile"
         create_git_repo
       end
 
       in_git_repo do
-        Shell.system "git checkout -b br"
-        Shell.system "echo 'blah' >> one"
-        Shell.system "git add one"
-        Shell.system "git commit -m 'commit on branch'"
-        Shell.system "git checkout master"
-        Shell.system "git merge --no-ff br"
+        RakeCommit::Shell.system "git checkout -b br"
+        RakeCommit::Shell.system "echo 'blah' >> one"
+        RakeCommit::Shell.system "git add one"
+        RakeCommit::Shell.system "git commit -m 'commit on branch'"
+        RakeCommit::Shell.system "git checkout master"
+        RakeCommit::Shell.system "git merge --no-ff br"
 
-        assert_equal 3, Shell.backtick("git log --pretty=oneline").split("\n").size
-        Shell.system "yes | ../../../bin/rake_commit --no-collapse"
+        assert_equal 3, RakeCommit::Shell.backtick("git log --pretty=oneline").split("\n").size
+        RakeCommit::Shell.system "yes | ../../../bin/rake_commit --no-collapse"
 
-        log_lines = Shell.backtick("git log --pretty=oneline").split("\n")
+        log_lines = RakeCommit::Shell.backtick("git log --pretty=oneline").split("\n")
         assert_equal 3, log_lines.size
         assert_match /Merge branch 'br'/, log_lines[0]
         assert_match /commit on branch/, log_lines[1]
@@ -138,16 +138,16 @@ class IntegrationTest < Test::Unit::TestCase
     Dir.chdir(TMP_DIR) do
       FileUtils.mkdir "git_repo"
       Dir.chdir("git_repo") do
-        Shell.system "touch Rakefile"
+        RakeCommit::Shell.system "touch Rakefile"
         create_git_repo
       end
 
       in_git_repo do
-        Shell.system "touch new_file"
-        Shell.system "git add new_file"
-        Shell.system "yes | ../../../bin/rake_commit --incremental"
+        RakeCommit::Shell.system "touch new_file"
+        RakeCommit::Shell.system "git add new_file"
+        RakeCommit::Shell.system "yes | ../../../bin/rake_commit --incremental"
 
-        log_lines = Shell.backtick("git log --pretty=oneline").split("\n")
+        log_lines = RakeCommit::Shell.backtick("git log --pretty=oneline").split("\n")
         assert_equal 2, log_lines.size
         assert_match /y - y/, log_lines.first
       end
@@ -158,13 +158,13 @@ class IntegrationTest < Test::Unit::TestCase
     Dir.chdir(TMP_DIR) do
       FileUtils.mkdir "git_repo"
       Dir.chdir("git_repo") do
-        Shell.system "touch Rakefile"
+        RakeCommit::Shell.system "touch Rakefile"
         create_git_repo
       end
 
       in_git_repo do
-        Shell.system "touch new_file"
-        fail_lines = Shell.backtick("yes | ../../../bin/rake_commit --incremental 2>&1", false).split("\n")
+        RakeCommit::Shell.system "touch new_file"
+        fail_lines = RakeCommit::Shell.backtick("yes | ../../../bin/rake_commit --incremental 2>&1", false).split("\n")
         assert_not_nil fail_lines.grep(/nothing added to commit but untracked files present/)
       end
     end
@@ -174,16 +174,16 @@ class IntegrationTest < Test::Unit::TestCase
     Dir.chdir(TMP_DIR) do
       FileUtils.mkdir "git_repo"
       Dir.chdir("git_repo") do
-        Shell.system "echo 'task :default do; end' >> Rakefile"
+        RakeCommit::Shell.system "echo 'task :default do; end' >> Rakefile"
         create_git_repo
       end
 
       in_git_repo do
-        Shell.system "git config user.name someone"
-        Shell.system "touch new_file"
-        Shell.system "yes | ../../../bin/rake_commit --without-prompt=pair"
+        RakeCommit::Shell.system "git config user.name someone"
+        RakeCommit::Shell.system "touch new_file"
+        RakeCommit::Shell.system "yes | ../../../bin/rake_commit --without-prompt=pair"
 
-        log_lines = Shell.backtick("git log | grep Author").split("\n")
+        log_lines = RakeCommit::Shell.backtick("git log | grep Author").split("\n")
         assert_match /\AAuthor: someone <.*>\z/, log_lines.first
       end
     end
@@ -193,18 +193,18 @@ class IntegrationTest < Test::Unit::TestCase
     Dir.chdir(TMP_DIR) do
       FileUtils.mkdir "git_repo"
       Dir.chdir("git_repo") do
-        Shell.system "echo 'task :default do; end' >> Rakefile"
+        RakeCommit::Shell.system "echo 'task :default do; end' >> Rakefile"
         create_git_repo
       end
 
       in_git_repo do
-        Shell.system "git config user.name someone"
-        Shell.system "touch new_file"
-        Shell.system "yes | ../../../bin/rake_commit --without-prompt=pair --without-prompt=feature"
+        RakeCommit::Shell.system "git config user.name someone"
+        RakeCommit::Shell.system "touch new_file"
+        RakeCommit::Shell.system "yes | ../../../bin/rake_commit --without-prompt=pair --without-prompt=feature"
 
-        log_lines = Shell.backtick("git log | grep Author").split("\n")
+        log_lines = RakeCommit::Shell.backtick("git log | grep Author").split("\n")
         assert_match /\AAuthor: someone <.*>\z/, log_lines.first
-        log_lines = Shell.backtick("git log --pretty=oneline").split("\n")
+        log_lines = RakeCommit::Shell.backtick("git log --pretty=oneline").split("\n")
         assert_match /\A\w+ y\z/, log_lines.first
       end
     end
@@ -214,15 +214,15 @@ class IntegrationTest < Test::Unit::TestCase
     Dir.chdir(TMP_DIR) do
       FileUtils.mkdir "git_repo"
       Dir.chdir("git_repo") do
-        Shell.system "echo 'task :default do; end' >> Rakefile"
+        RakeCommit::Shell.system "echo 'task :default do; end' >> Rakefile"
         create_git_repo
       end
 
       in_git_repo do
-        Shell.system "touch new_file"
-        Shell.system "yes | ../../../bin/rake_commit --without-prompt=feature"
+        RakeCommit::Shell.system "touch new_file"
+        RakeCommit::Shell.system "yes | ../../../bin/rake_commit --without-prompt=feature"
 
-        log_lines = Shell.backtick("git log --pretty=oneline").split("\n")
+        log_lines = RakeCommit::Shell.backtick("git log --pretty=oneline").split("\n")
         assert_match /\A\w+ y\z/, log_lines.first
       end
     end
@@ -232,30 +232,30 @@ class IntegrationTest < Test::Unit::TestCase
     Dir.chdir(TMP_DIR) do
       FileUtils.mkdir "git_repo"
       Dir.chdir("git_repo") do
-        Shell.system "echo 'task :default do; end' >> Rakefile"
+        RakeCommit::Shell.system "echo 'task :default do; end' >> Rakefile"
         create_git_repo
       end
 
       in_git_repo do
-        Shell.system "echo '--without-prompt=feature' > .rake_commit"
-        Shell.system "yes | ../../../bin/rake_commit"
+        RakeCommit::Shell.system "echo '--without-prompt=feature' > .rake_commit"
+        RakeCommit::Shell.system "yes | ../../../bin/rake_commit"
 
-        log_lines = Shell.backtick("git log --pretty=oneline").split("\n")
+        log_lines = RakeCommit::Shell.backtick("git log --pretty=oneline").split("\n")
         assert_match /\A\w+ y\z/, log_lines.first
       end
     end
   end
 
   def create_git_repo
-    Shell.system "git init"
-    Shell.system "git add Rakefile"
-    Shell.system "git commit -m 'Added Rakefile'"
-    Shell.system "git checkout -b not_master"
+    RakeCommit::Shell.system "git init"
+    RakeCommit::Shell.system "git add Rakefile"
+    RakeCommit::Shell.system "git commit -m 'Added Rakefile'"
+    RakeCommit::Shell.system "git checkout -b not_master"
     sleep 1 # Ensure that the first commit is at least one second older
   end
 
   def in_git_repo(&block)
-    Shell.system "git clone file://#{TMP_DIR}/git_repo git_wc"
+    RakeCommit::Shell.system "git clone file://#{TMP_DIR}/git_repo git_wc"
     Dir.chdir("git_wc") do
       yield
     end
