+++ test/test_helper.rb
@@ -4,7 +4,7 @@ unless defined?(TEST_HELPER_LOADED)
   require "rubygems"
   require "rake"
   Dir.glob(File.dirname(__FILE__) + "/../lib/tasks/**/*.rake").each { |rakefile| load rakefile }
-  require File.dirname(__FILE__) + "/../lib/commit"
+  require File.dirname(__FILE__) + "/../lib/rake_commit"
 
   require "test/unit"
   require "mocha"
